refactor: Move Appointment model to its own file
Moved the Appointment model from ApplicationDbContext.cs to its own file, Appointment.cs, to improve code organization and maintainability. This change follows the single responsibility principle and makes the codebase easier to navigate.
